refactor: simplify TradeBasicConfirmDetails
