+++ apps/cowswap-frontend/src/modules/trade/utils/getLimitPriceFromReceiveAmount.ts
@@ -1,0 +1,16 @@
+import { FractionUtils } from '@cowprotocol/common-utils'
+import { Price, Currency } from '@uniswap/sdk-core'
+
+import { ReceiveAmountInfo } from '../types'
+
+export function getLimitPriceFromReceiveAmount({ afterNetworkCosts, afterSlippage }: ReceiveAmountInfo): Price<Currency, Currency> | null {
+
+  const quoteAmount = FractionUtils.amountToAtLeastOneWei(afterSlippage.buyAmount)
+
+  if (!quoteAmount) return null
+
+  return new Price({
+    quoteAmount,
+    baseAmount: afterNetworkCosts.sellAmount,
+  })
+}
